POM-32. Use Curator entries from the root pom.
review: https://review.kiji.org/r/1798
issue: https://jira.kiji.org/POM-32

SCHEMA-741. Tests should cleanup dummy instances when using an external HBase

issue: https://jira.kiji.org/browse/SCHEMA-741
review: https://review.kiji.org/r/1796

SCHEMA-677. consolidate layout tracking in Kiji layer

issue: https://jira.kiji.org/browse/SCHEMA-677
review: https://review.kiji.org/r/1786

SCHEMA-771. Create a unique user ID per TableLayoutMonitor

issue: https://jira.kiji.org/browse/SCHEMA-771
review: https://review.kiji.org/r/1835

SCHEMA-752. switch to Curator client for ZooKeeper interactions

review: https://review.kiji.org/r/1803
issue: https://jira.kiji.org/SCHEMA-752

SCHEMA-774. Fix race condition in test

issue: https://jira.kiji.org/browse/SCHEMA-774
review: https://review.kiji.org/r/1837

    SCHEMA-644. Foo Table layout specifies an avro type that doesn't exist.          issue: https://jira.kiji.org/browse/SCHEMA-644

    SCHEMA-644. Foo Table layout specifies an avro type that doesn't exist.

    issue: https://jira.kiji.org/browse/SCHEMA-644

review: https://review.kiji.org/r/1777
contributed-by: akahn (adam kahn, akahn@wibidata.com)

Update versions for release

[maven-release-plugin] prepare release kiji-schema-root-1.4.1

[maven-release-plugin] prepare for next development iteration

SCHEMA-774. Fix race condition in test

issue: https://jira.kiji.org/browse/SCHEMA-774
review: https://review.kiji.org/r/1840

SCHEMA-764. Update REVIEWERS file.

issue: https://jira.kiji.org/browse/SCHEMA-764
review: https://review.kiji.org/r/1823

SCHEMA-769. Add a KijiResultScanner implementation

issue: https://jira.kiji.org/browse/SCHEMA-769
review: https://review.kiji.org/r/1834

SCHEMA-777. HBaseKijiResult returns incorrect number of values when page size is not a multiple of max versions.

issue: https://jira.kiji.org/browse/SCHEMA-777
review: https://review.kiji.org/r/1849

SCHEMA-775. Refactor KijiColumnNameTranslator/HBaseColumnName/LayoutCapsule

issue: https://jira.kiji.org/browse/SCHEMA-775
review: https://review.kiji.org/r/1838

review: https://review.kiji.org/r/1838
contributed-by: dan (Dan Burkert, dan@wibidata.com)

+++ kiji-schema/src/test/java/org/kiji/schema/TestHBaseKijiRowData.java
@@ -56,7 +56,7 @@ import org.kiji.schema.impl.LayoutCapsule;
 import org.kiji.schema.impl.hbase.HBaseKijiRowData;
 import org.kiji.schema.impl.hbase.HBaseKijiTable;
 import org.kiji.schema.layout.CellSpec;
-import org.kiji.schema.layout.KijiColumnNameTranslator;
+import org.kiji.schema.layout.HBaseColumnNameTranslator;
 import org.kiji.schema.layout.KijiTableLayouts;
 import org.kiji.schema.layout.TranslatedColumnName;
 import org.kiji.schema.layout.impl.LayoutCapsule;
@@ -217,7 +217,7 @@ public class TestHBaseKijiRowData extends KijiClientTest {
 
     HBaseKijiRowData input = new HBaseKijiRowData(mTable, dataRequest, row0, result, null);
     input.getMap();
-    final int integer = (Integer) input.getMostRecentValue("family", "qual3");
+    final int integer = input.getMostRecentValue("family", "qual3");
     assertEquals(42, integer);
   }
